move consumables into expansion folder
